ml/requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
tensorflow==2.16.1
joblib==1.4.2
firebase-admin==6.5.0
scipy==1.13.0
matplotlib==3.9.0
seaborn==0.13.2
google-cloud-firestore==2.16.0
